llm-proxy/pyproject.toml
[project]
name = "codelab-llm-proxy"
version = "0.1.0"
description = "CodeLab LLM Proxy Service - Унифицированный доступ к LLM провайдерам"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Sergey Penkovsky", email = "sergey.penkovsky@gmail.com" }]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.1",
    "pydantic>=2.5.1",
    "openai>=1.3.5",
    "anthropic>=0.7.4",
    "tenacity>=8.2.3",
    "sse-starlette>=1.6.5",
    "openai>=1.3.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = ["ruff", "ty", "pytest", "pytest-asyncio", "pytest-cov"]

[dependency-groups]
dev = [
    "ruff>=0.14.8",
    "ty>=0.0.1a31",
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
]

[tool.pytest]
testpaths = ["tests"]
python_files = ["test_*.py"]

#[tool.ty]
#src = "pyproject.toml"
#out = "requirements.txt"
#
#[tool.ty.rules]
#unused-ignore-comment      = "warn"   # Предупреждать, если ignore-комментарий не используется
#redundant-cast            = "ignore" # Не сообщать о лишних cast()
#possibly-missing-attribute= "error"  # Всегда считать ошибкой отсутствие атрибута
#possibly-missing-import   = "error"  # Всегда считать ошибкой отсутствие зависимости
#unresolved-import         = "ignore" # Игнорировать неразрешённые импорты (иногда нужно для динамических импортов, которые намеренно не всегда присутствуют)
